Fix syntax of two programs (powint and binarysearch) and add powint-erroneous (#71)
* Fix syntax of powint.dana

Use"=" instead of "==" for equality and moved "var temp is int"
declaration to the beginning of the block. The program still contains a
semantic error so I copied it to the dana/programs-erroneous directory.

* Fix syntax in binarysearch program

Adjusted the bAdded a missing comma in a procedure call, changed
function calls to use () instead of :, specifed a fixed size for the
local array variable declaration, and corrected the assignment operator
from "=" to ":=".